divisão do componente em uma pasta específica + indentação
