+++ src/app/(Site)/(LandingPage)/page.tsx
@@ -1,4 +1,4 @@
-import PaqHeader from "@/app/components/landing/PaqHeader";
+import PaqHeader from "@/app/components/Header/PaqHeader";
 import ConectarETransformar from "@/app/components/ConectarETransformar";
 import UmFuturoInovadorDiversoComecaAqui from "@/app/components/UmFuturoInovadorDiversoComecaAqui";
 import PaqFaq from "@/app/components/PaqFaq";
